doesnotwor
Perhaps doing this with minimal rebuilds is not really possible with
dune yet.

We should instead try to write something suitable in a CI, e.g. to be
run at the end of a build when you have the cmt files anyway.
(or parse dune rules like before, or use glob_files_rec with cmt?)